Reorganized entire project structure and split css files into individual files for each site
