+++ src/front/components/characterCreation/ClassSelector.jsx
@@ -1,11 +1,11 @@
-import useGlobalReducer from "../hooks/useGlobalReducer.jsx";
+import useGlobalReducer from "../../hooks/useGlobalReducer";
 
 export const ClassSelector = ({ setCharacterClass }) => {
     const { store, dispatch } = useGlobalReducer()
     return (
         <div>
             <label className="input-selector mt-3" htmlFor="name">SELECT A CLASS</label>
-            <div className="race-selector-input mt-3">
+            <div className="grid-selector-input mt-3">
                 {
                     store.classes
                         ? store.classes.map((characterClass, index) => {
