+++ src/front/components/characterCreation/NameSelector.jsx
