+++ src/front/components/characterSite/CharacterInfoBlock.jsx
