+++ src/front/components/characterSite/ConsumablesTable.jsx
