+++ src/front/components/characterSite/DiceBar.jsx
