+++ src/front/components/characterSite/SkillsTable.jsx
