+++ src/front/components/characterSite/StatBlock.jsx
