+++ src/front/components/signUpLogIn/SignUpLogInInterface.jsx
@@ -1,22 +1,22 @@
 import React, { useEffect, useState } from "react"
 import { Link } from "react-router-dom"
-import useGlobalReducer from "../hooks/useGlobalReducer.jsx";
+import useGlobalReducer from "../../hooks/useGlobalReducer.jsx";
 
 const backendUrl = import.meta.env.VITE_BACKEND_URL
 
 
 export default function InputForm() {
-    const { store, dispatch } = useGlobalReducer()
-    const [isLogin, setIsLogin] = useState(true);
-    const [formInput, setFormInput] = useState({email: "", password: ""});
+  const { store, dispatch } = useGlobalReducer()
+  const [isLogin, setIsLogin] = useState(true);
+  const [formInput, setFormInput] = useState({ email: "", password: "" });
 
-    const handleInput = (e) => 
-        setFormInput({ ...formInput, [e.target.name]: e.target.value });
+  const handleInput = (e) =>
+    setFormInput({ ...formInput, [e.target.name]: e.target.value });
 
-    const handleSubmit = async (e) => {
+  const handleSubmit = async (e) => {
     e.preventDefault();
 
-    const endpoint = isLogin ? "/api/login" : "/api/signup"; 
+    const endpoint = isLogin ? "/api/login" : "/api/signup";
     const response = await fetch(endpoint, {
       method: "POST",
       headers: { "Content-Type": "application/json" },
@@ -27,7 +27,7 @@ export default function InputForm() {
     console.log(data);
   };
 
-    return (
+  return (
     <div className="form-container">
       <h2>{isLogin ? "LOG IN" : "SIGN UP"}</h2>
       <form onSubmit={handleSubmit}>
@@ -35,17 +35,17 @@ export default function InputForm() {
         <input
           type="email" name="email" placeholder="Enter your email..." onChange={handleInput} value={formInput.email} required
         />
-       
+
         <input
           type="password" name="password" placeholder="Enter your password..." onChange={handleInput} value={formInput.password} required
         />
-        
-        <button type="submit">SUBMIT</button> 
+
+        <button type="submit">SUBMIT</button>
 
       </form>
-    {/* <LogInButton /> this will need to be properly routed and linked to the top right of the NavBar but it will need to be present here to switch. */}
+      {/* <LogInButton /> this will need to be properly routed and linked to the top right of the NavBar but it will need to be present here to switch. */}
     </div>
-    
+
   );
 }
