+++ src/front/components/spellSite/SpellBasicInfo.jsx
