+++ src/front/components/spellSite/SpellBonusInfo.jsx
