--- src/front/index.css
@@ -1,128 +1,0 @@
-/* GENERAL SITE */
-body {
-  background-color: #40202c;
-}
-.header-primary {
-  background-color: #8c373e;
-}
-.header-secondary {
-  background-color: #bfb699;
-}
-.accent {
-  color: #008080;
-}
-.col {
-  text-align: center;
-  justify-content: center;
-}
-.row {
-  text-align: center;
-  justify-content: center;
-}
-.info-box {
-  background-color: #8c373e;
-  border-style: solid;
-  border-color: #bfb699;
-  border-width: 15px;
-}
-/* CHARACTER INFO PAGE SPECIFIC */
-.character-info-box {
-  background-color: whitesmoke;
-  color: #008080;
-  border-style: solid;
-  border-color: #bfb699;
-  border-width: 5px;
-}
-.character-picture {
-  border-style: solid;
-  border-color: #bfb699;
-  border-width: 5px;
-  border-radius: 15px 50px;
-}
-/* STAT SPECIFIC */
-.stat-block-box {
-  background-color: #bfb699;
-  border-style: solid;
-  border-color: #008080;
-  border-width: 5px;
-}
-.stat-block-info {
-  background-color: whitesmoke;
-  color: #008080;
-}
-/* DICE BAR SPECIFIC */
-.dice-block {
-  background-color: #bfb699;
-  border-style: solid;
-  border-color: #008080;
-  color: #008080;
-  border-width: 5px;
-  aspect-ratio: 1 / 1;
-  width: 75%;
-  margin: 5px;
-}
-.dice-box {
-  margin: auto;
-  width: 50%;
-}
-table {
-  text-align: left;
-}
-/* SPELL LIST SPECIFIC */
-.spell-name {
-  border-bottom: solid;
-  border-color: #bfb699;
-  border-width: 15px;
-}
-.spells-box {
-  width: 20%;
-}
-.spells-options {
-  aspect-ratio: 1 / 1;
-  overflow: scroll;
-}
-.spell-information-box {
-
-/* Login/Signup specific */
-.form-container{
-  background-color: #8c373e;
-  border-style: solid;
-  border-color: #bfb699;
-  border-width: 15px;
-/* CHARACTER CREATOR PAGE SPECIFIC */
-.input-selector {
-  width: 15%;
-  margin-inline: 5px;
-  background-color: whitesmoke;
-  color: #008080;
-  border-style: solid;
-  border-color: #bfb699;
-  border-width: 5px;
-}
-.input-field {
-  width: 25%;
-  margin-inline: 5px;
-  background-color: whitesmoke;
-  color: #008080;
-  border-style: solid;
-  border-color: #40202c;
-  border-width: 5px;
-}
-.race-selector-input {
-  column-count: 3;
-}
-.input-selection {
-  width: 100%;
-  background-color: whitesmoke;
-  color: #008080;
-  border-style: solid;
-  border-color: #bfb699;
-  border-width: 5px;
-  overflow: scroll;
-  aspect-ratio: 2 / 1;
-}
-.spell-info-table table,
-td,
-tr {
-  border: 1px solid;
-}
