+++ src/styles/characterSite.css
@@ -1,0 +1,43 @@
+/* CHARACTER INFO PAGE SPECIFIC */
+.character-info-box {
+  background-color: whitesmoke;
+  color: #008080;
+  border-style: solid;
+  border-color: #bfb699;
+  border-width: 5px;
+}
+.character-picture {
+  border-style: solid;
+  border-color: #bfb699;
+  border-width: 5px;
+  border-radius: 15px 50px;
+}
+/* STAT SPECIFIC */
+.stat-block-box {
+  background-color: #bfb699;
+  border-style: solid;
+  border-color: #008080;
+  border-width: 5px;
+}
+.stat-block-info {
+  background-color: whitesmoke;
+  color: #008080;
+}
+/* DICE BAR SPECIFIC */
+.dice-block {
+  background-color: #bfb699;
+  border-style: solid;
+  border-color: #008080;
+  color: #008080;
+  border-width: 5px;
+  aspect-ratio: 1 / 1;
+  width: 75%;
+  margin: 5px;
+}
+.dice-box {
+  margin: auto;
+  width: 50%;
+}
+table {
+  text-align: left;
+}
